feat: add `from-tokenn` command to build deployments from on-chain tokens
