+++ docs/api/cli/build_images.md
@@ -1,0 +1,40 @@
+<a id="autonomy.cli.build_images"></a>
+
+# autonomy.cli.build`_`images
+
+Build images.
+
+<a id="autonomy.cli.build_images.build_images"></a>
+
+#### build`_`images
+
+```python
+@click.command(name="build-images")
+@click.option(
+    "--build-dir",
+    type=click.Path(dir_okay=True),
+    help="Path to build dir.",
+)
+@click.option(
+    "--packages-dir",
+    type=click.Path(dir_okay=True),
+    help="Path to packages folder (for local usage).",
+)
+@click.option(
+    "--skaffold-dir",
+    type=click.Path(exists=True, dir_okay=True),
+    help="Path to directory containing the skaffold config.",
+)
+@click.option(
+    "--version",
+    type=str,
+    default=DEFAULT_IMAGE_VERSION,
+    help="Image version.",
+)
+@click.option("--push", is_flag=True, default=False, help="Push image after build.")
+@image_profile_flag()
+def build_images(profile: str, packages_dir: Optional[Path], build_dir: Optional[Path], skaffold_dir: Optional[Path], version: str, push: bool) -> None
+```
+
+Build image using skaffold.
+
